refactor dot_writer and create_dot
