+++ src/output/dot_writer.cpp
@@ -1,6 +1,70 @@
 #include "dot_writer.h"
 
 
+void Dot_writer::read_data(AllData& alldata) {
+
+    std::stack<Node*> parent_nodes;
+
+    int call_id = 0;
+
+    for ( const auto& region : alldata.call_path_tree ) {
+
+        // std::shared_ptr<Node> node = std::make_shared<Node>();
+        Node* node = new Node;
+
+        node->call_id = call_id;
+        ++call_id;
+
+        std::string region_name = alldata.definitions.regions.get(region.function_id)->name;
+        node->region = region_name;
+
+        node->num_children = region.children.size();
+
+        // mpi time resolution
+        double timerResolution = (double)alldata.metaData.timerResolution;
+
+        // accumulate data over all locations
+        for (const auto& location : region.node_data) {
+            // filter per rank
+            if(alldata.params.rank == -1 || alldata.params.rank == location.first){
+
+                node->invocations += location.second.f_data.count;
+
+                double incl_time = location.second.f_data.incl_time / timerResolution;
+                if (node->min_incl_time > incl_time)
+                    node->min_incl_time = incl_time;
+                if (node->max_incl_time < incl_time)
+                    node->max_incl_time = incl_time;
+                node->sum_incl_time += incl_time;
+
+                double excl_time = location.second.f_data.excl_time / timerResolution;
+                if (node->min_excl_time > excl_time)
+                    node->min_excl_time = excl_time;
+                if (node->max_excl_time < excl_time)
+                    node->max_excl_time = excl_time;
+                node->sum_excl_time += excl_time;
+            }
+        }
+
+        // average time
+        node->avg_incl_time = node->sum_incl_time / region.node_data.size();
+        node->avg_excl_time = node->sum_excl_time / region.node_data.size();
+
+        // set parent <-> child relationship
+        if(parent_nodes.size() != 0){
+
+            parent_nodes.top()->children.push_back(node);
+
+            node->parent = parent_nodes.top();
+            parent_nodes.pop();
+        }
+
+        for( int i = 0; i < node->num_children; ++i )
+            parent_nodes.push(node);
+        nodes.push_back(node);
+    }
+}
+
 bool Dot_writer::open(){
     std::string filename = params.output_file_prefix + ".dot";
     result_file.open(filename);
@@ -10,7 +74,7 @@ bool Dot_writer::open(){
         result_file
         << "digraph call_tree {\n"
         << "graph [splines=" << splines <<", ranksep=" << ranksep << "];\n"
-        << "node [shape = "<< shape << ", colorscheme="<< colorscheme << "];\n"
+        << "node [shape = " << shape << ", colorscheme="<< colorscheme << "];\n"
         << "edge [];\n"
         << std::endl;
 
@@ -26,6 +90,64 @@ bool Dot_writer::open(){
 void Dot_writer::close(){
     result_file << "}" << std::endl;
     result_file.close();
+}
+
+void Dot_writer::print(){
+
+    top_nodes();
+    filter_nodes();
+
+    gather_time_data();
+
+    for( auto& node : nodes ){
+        if(node->state != NodeState::dontprint)
+            print_node(*node);
+    }
+}
+
+void Dot_writer::print_node(Node& node){
+    result_file
+    << "\""             << node.call_id       << "\" [\n"
+    << " label = \""    << node.region        << "\\l\n";
+
+    if(node.state == NodeState::full){
+        result_file
+        << " invocations: " << node.invocations   << "\\l\n"
+        << " include time:"                       << "\\l\n"
+        << "  min: "        << node.min_incl_time << "\\l\n"
+        << "  max: "        << node.max_incl_time << "\\l\n"
+        << "  sum: "        << node.sum_incl_time << "\\l\n"
+        << "  avg: "        << node.avg_incl_time << "\\l\n"
+        << " exclude time:"                       << "\\l\n"
+        << "  min: "        << node.min_excl_time << "\\l\n"
+        << "  max: "        << node.max_excl_time << "\\l\n"
+        << "  sum: "        << node.sum_excl_time << "\\l\n"
+        << "  avg: "        << node.avg_excl_time << "\\l\n";
+    }
+
+    result_file << " \"\n";
+
+    // colorize node
+    if(node.state == NodeState::full){
+        result_file << " fillcolor=" << get_node_color(node.sum_excl_time) << ",\n";
+        result_file << " style=filled\n";
+    }
+
+    // closing node tag
+    result_file << "];" << std::endl;
+
+    // set edge between this node and its parent
+    if( node.parent){
+        result_file
+        << node.parent->call_id
+        << " -> "
+        << node.call_id
+        << ";\n"
+        << std::endl;
+    }
+
+    // mark as printed
+    node.state = NodeState::printed;
 }
 
 void Dot_writer::filter_nodes(){
@@ -58,52 +180,6 @@ void Dot_writer::filter_nodes(){
     }
 }
 
-void Dot_writer::print_node(Node& node){
-    result_file
-    << "\""             << node.call_id       << "\" [\n"
-    << " label = \""    << node.region        << "\\l\n";
-
-    if(node.state == NodeState::full){
-        result_file
-        << " invocations: " << node.invocations   << "\\l\n"
-        << " include time:"                       << "\\l\n"
-        << "  min: "        << node.min_incl_time << "\\l\n"
-        << "  max: "        << node.max_incl_time << "\\l\n"
-        << "  sum: "        << node.sum_incl_time << "\\l\n"
-        << "  avg: "        << node.avg_incl_time << "\\l\n"
-        << " exclude time:"                       << "\\l\n"
-        << "  min: "        << node.min_excl_time << "\\l\n"
-        << "  max: "        << node.max_excl_time << "\\l\n"
-        << "  sum: "        << node.sum_excl_time << "\\l\n"
-        << "  avg: "        << node.avg_excl_time << "\\l\n";
-    }
-
-    result_file << " \"\n";
-
-    // colorize node
-    if(node.state == NodeState::full){
-        result_file << " fillcolor=" << get_node_color(node.sum_excl_time) << ",\n";
-        result_file << " style=filled\n";
-    }
-
-
-    // closing node tag
-    result_file << "];" << std::endl;
-
-    // set edge between this node and its parent
-    if( node.parent){
-        result_file
-        << node.parent->call_id
-        << " -> "
-        << node.call_id
-        << ";\n"
-        << std::endl;
-    }
-
-    // mark as printed
-    node.state = NodeState::printed;
-}
-
 void Dot_writer::gather_time_data(){
     for( const auto& node : nodes ){
 
@@ -118,13 +194,13 @@ void Dot_writer::gather_time_data(){
 
         total_time += node->sum_excl_time;
     }
-
-    timerange = max_time - min_time;
 }
 
 void Dot_writer::mark_predecessors(Node* node){
+    // void Dot_writer::mark_predecessors(std::shared_ptr<Node> node){
 
     Node* curr = node;
+    // std::shared_ptr<Node> curr = node;
 
     while( curr->parent != nullptr ){
         curr = curr->parent;
@@ -138,12 +214,13 @@ void Dot_writer::mark_predecessors(Node* node){
 
 void Dot_writer::top_nodes(){
 
-    if(num == 0)
+    if(params.top_nodes == 0)
         return;
 
     filter = true;
 
     std::sort(nodes.begin(), nodes.end(), [](Node* a, Node* b){
+    // std::sort(nodes.begin(), nodes.end(), [](std::shared_ptr<Node> a, std::shared_ptr<Node> b){
         return a->sum_excl_time > b->sum_excl_time;
     });
 
@@ -158,6 +235,7 @@ void Dot_writer::top_nodes(){
 }
 
 int Dot_writer::get_node_color(const double time){
+    double timerange = max_time - min_time;
     int color_code = num_colors;
     for( int i = 0; i < num_colors; ++i ){
         if( time >= timerange/num_colors*i+min_time )
@@ -166,19 +244,11 @@ int Dot_writer::get_node_color(const double time){
     return color_code;
 }
 
-void Dot_writer::add_node(Node* node){
-    nodes.push_back(node);
-}
+// void Dot_writer::add_node(Node* node){
+//     // void Dot_writer::add_node(std::shared_ptr<Node> node){
+//     nodes.push_back(node);
+// }
 
-void Dot_writer::print(){
-
-    top_nodes();
-    filter_nodes();
-
-    gather_time_data();
-
-    for( auto& node : nodes ){
-        if(node->state != NodeState::dontprint)
-            print_node(*node);
-    }
-}
+// void Dot_writer::add_nodes(std::vector<Node*>& nodes){
+//     this->nodes = nodes;
+// }
